ID-Verification-UiPathPackage-v2/requirements.txt
numpy==1.21.2
opencv-python==4.5.5.62
mtcnn==0.1.1
arcface==0.0.8
pillow==8.3.1
astropy==5.0.1